requirements.txt
aiofiles==23.2.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.42.1
fsspec==2023.9.0
gradio==3.44.0
gradio_client==0.5.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.1
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.7.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
orjson==3.9.7
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.3
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tiktoken==0.3.3
toolz==0.12.0
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
websockets==11.0.3
